requirements.txt
python-dotenv==1.0.0
pdfminer.six==20221105
langchain==0.0.350
langchain-community==0.0.10
scikit-learn==1.3.2
numpy==1.26.2
streamlit==1.29.0
PyPDF2==3.0.1
ollama==0.1.7